requirements.txt
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==7.1.1
crashtest==0.3.1
cryptography==38.0.1
Deprecated==1.2.13
distlib==0.3.6
docopt==0.6.2
dulwich==0.20.46
filelock==3.8.0
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==3.4
itsdangerous==1.1.0
jaraco.classes==3.2.3
Jinja2==2.11.1
jsonschema==4.16.0
jwcrypto==1.4.2
keyring==23.9.3
line-bot-sdk==1.8.0
lockfile==0.12.2
MarkupSafe==1.1.1
more-itertools==8.14.0
msgpack==1.0.4
packaging==21.3
pexpect==4.8.0
pipreqs==0.4.11
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.1.2
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.5.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
shellingham==1.5.0
six==1.16.0
tomlkit==0.11.5
urllib3==1.26.12
virtualenv==20.16.5
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.14.1
xattr==0.9.9
yarg==0.1.9
